Bump triton to 3.5.x, aiter, and apex (#2728)
PyTorch updated pybind11 to v3 and removed the
use of `_PYBIND11_COMPILER_TYPE`. All the downstream projects now need
to remove the use of `_PYBIND11_COMPILER_TYPE`. More specifically, we
needed an aiter that removed this. Bumping aiter submodule in PyTorch
and apex commit which also has aiter as a submodule.

Bumping triton as well to 3.5.x

Wheel Validation:
http://rocm-ci.amd.com/view/preview/job/pytorch-latest-manylinux-wheels-preview/177/